+++ src/pixi-native/video/index.ts
@@ -1,3 +1,2 @@
 export * from "./NativeVideo.ts";
-export * from "./NativeVideoSprite.ts";
 export * from "./nv12ToRgba.ts";